incorperate latest blessings changes
this is more or less the version submitted to erik, with glues to make the API compatible with the way the BBS uses it. If the changes ever get merged upstream, the "userland" bbs should use "term.inkey()", and "term.center()" etc. directly, which will pair down LOC significantly

the reason this change was pulled in, was tintin++ client will send '\x1b[A' in parts, and suprisngly the server never implemented an escape delay -- but this was implemented upstream in blessings, so instead of re-writing yet another esc_delay() routine, we simply pulled in blessings early to fix this bug