Allow 'anonymous' login handles to be defined
Previously, the handle 'anonymous' was hardcoded
as the "anonymous user" which would be toggled
by another command line option.

Now, the handles that have meaning as 'anonymous'
may be configured. I intend to use this for good
fun by making the handle 'root' an anonymous user
so those fucking Chinese netscanners have something
interesting to show in their logs.