Move code from engine.py to fail2ban.py
also make fail2ban a function wrapper -- call it
only one time to configure it -- it now returns
a function that can be called at on-connect time.

The function accept() now accepts third argument,
check_ban, which is a function to be called, and
is usually provided by get_fail2ban_function(),

-- if fail2ban is not configured (section
'fail2ban', key 'enabled' is not True), then an
anonymous function is returned that always
returns True.