Tetris and Weather Fixes
Summary tetris.py
echo(syncterm_setfont('cp437'))  to
echo_unbuffered(syncterm_setfont('cp437'))  Original buffered settings
makes this command output after the ANSI screen is drawn. It should push
out prior to displaying the ANSI screen.

Summary weather.py
1. loop was not saving new searches or prompting the user to save
because it was outside of the loop.  This affected current users since
it was not updating and only retrieving their last save.  And affected
new users because it would never save a new location.

2. Once this was fixed I noticed Accu Weather API doesn't return postal
code anymore on findcity.  This is causing the get.location to retrieve
a bad value on postal lookup for pre-populating the search field.
Changing this to City, State lookup solves the issues and looks nice
too.